requirements.txt
torch>=1.8.0
torchvision>=0.9.0
datasets>=1.18.0
PyYAML>=5.4.1
tqdm>=4.62.3
tensorboard>=2.12.0 